refactor: 로또 번호를 구입하는 함수를 유틸로 분리
